server: fix AbortError from finish/TapParser.end
Previously we used a ReadableStream, and fed lines submitted by
browsoers (in handleTap), to the readableControllerParam. Then,
we piped this to tapParser (using `readable.pipeTo(stream.Writable.toWeb(tapParser));`).

In the CLI mode, we run process.exit() as soon as we're done so it
works fine, but in programmatic mode, I noticed that there is actually
a late error being thrown:

```
AbortError: The operation was aborted
    at Parser.<anonymous> (node:internal/webstreams/adapters:205:22)
    at Parser.<anonymous> (node:internal/util:544:20)
    at Parser.onfinish (node:internal/streams/end-of-stream:112:16)
    at Parser.emit (node:events:507:28)
    at Parser.emitComplete (file:///qtap/node_modules/tap-parser/dist/esm/index.js:403:18)
    at Parser.end (file:///qtap/node_modules/tap-parser/dist/esm/index.js:383:14)
    at finish (file:///qtap/src/tap-finished.js:61:16)
    at check (file:///qtap/src/tap-finished.js:67:5)
    at Parser.<anonymous> (file:///qtap/src/tap-finished.js:79:5)
    at Parser.emit (node:events:507:28) {
  code: 'ABORT_ERR'
```

It's not clear why, but from a quick search I found
https://github.com/nodejs/help/issues/2412 which says that one
should not emit `finish` yourself, as tap-parser currently does
from `Parser.end()`.

Workaround this by using a simple callback (tapParser.write) as our
interface between handleTap and TapParser, instead of a ReadableStream.
This is much simpler anyway and is what I should have done in the first
place.