When serving test file form static server, use its actual path
Reason 1:

Instead of always proxying to `/` and always adding `<base href>`,
for the common case of passing a static file, serve it from its
natural URL. I didn't want to do this originally so that we only
have one way of doing this, but when trying it out on WordPress
I noticed that it relied on location.href for something other than
just fetching resources, it actually looks at it directly for words
it contains as well (for the password strength unit test).

This is likely not the only project that does something like this,
and arguably it is "surprising" to see `qtap test/index.html` served
from a URL other than `localhost:X/test/index.html`, which is how
any other test runner would serve if if it didn't need a proxy. So
in the interest of less surprise and making the proxy more transparent,
let's prefer to present on a more natural URL when possible.

Reason 2:

For external URL, I completely forgot about query parameters, which
should be visible client-side as well.