Remove 'stderr' endpoint in favor of sending stderr as TAP comment
Follows-up f75d8a1953.

This has the benefit of making it easy to guruantee a consistent order
between test results and stderr messages.

This is going to be useful in an upcoming change, which is to report
any stderr collected "during a test" along with test diagnostic.

This is difficult to do when there are likely race conditions, such
as when the two streams are separate.