cli: Improve help by changing browser from `["detect"]` to `"default"`
== Why ==

The array brackets add noise and needless specificity about an
implementation (namely that each -b value is collected in an array)
that doesn't matter to the end-user.

It also creates ambiguity over what might happen to the default when
you specify you own values. It is reasonable to assume that your
values add to the default, given that that is often how arrays are
used. And indeed, that is how Commander works by default (our parseArg
function overrides this so that the first value replaces the default).

By specifying a string default, it is obvious that when you set the
-b option yourself, it replaces the default value, since that is how
options generally work. Given an option `--foo` that defaults to X,
if you set `-foo Y`, then Y is used instead of X.

The fact that you may specify -b multiple times can then be conceived
of separately.

== Change ==

Before

```
-b, --browser <name>
        One or more browser names.
        Available: detect, firefox, chrome, …. (default: ["detect"])
```

After

```
-b, --browser <name>
        One or more browser names.
        Available: detect, firefox, chrome, …. (default: "detect")
```

== How ==

This commit could be a oneline change like so:

```
- const optionBrowserDefault = ['detect'];
+ const optionBrowserDefault = 'detect';
```

The `option(,,parseArg)` function allows the type to change between
defaultValue and parsed value, and also between two parses.
Our code in `qtap.run()` also accepts both string and array.

But, the TypeScript definition for Commander introduces an arbitrary
restriction on this function, demanding the type of the `prev` (list)
argument to match the returned type. I'm sure this helps some people
catch genuine mistakes, but here it's just in the way. We could slap
a `@ts-ignore` on this.

Another way of thinking about is that we just want to change the
help text (defaultValueDescription), and leave the actual value
as a consistent array. This has appeal and merit, too, but has the
downside of requring the entire option to be de-sugared because the
defaultValueDescription param isn't exposed to `program.option()`.