server: Support query paramters on local files
A common reason to set up a custom web server is to set query parameters
like "filter" or "module" for QUnit, or "grep" for Mocha.

By supporting passing query parameters to the built-in web servers
we remove a common need to set up a custom web server and make QTap
easier to use.