server: Allow `--debug` mode to reload the browser window
To ease debugging, allow the landing page to be loaded multiple times.

Normally, to catch mistakes and bugs, we do not permit this since we
would have anywhere to report on a second or third run of test results
from the same clientId, since we will have closed that stream and
finished the report on that client already.